fix avatar

+++ app/Http/Controllers/ClientApp/Api/v1/PostcardController.php
@@ -305,13 +305,16 @@ WHERE res.user_id <> ? or (user_id = ? and start is NULL)
 
     public function saveAvatar(Request $request)
     {
-        $link = $this->saveMediaContent($request->file('file'), 'user/'.$request->input('postcard_id').'/avatar', $request->input('media_content_type'));
-        Storage::disk('public')->delete(\Auth::user()->avatar);
+        $link = $this->saveMediaContent($request->file('file'), 'user/'.\Auth::user()->id.'/avatar', $request->input('media_content_type'));
+        if (\Auth::user()->avatar) {
+            Storage::disk('public')->delete(\Auth::user()->avatar);
+        }
+
         $mediaContent = \Auth::user()->update([
                 'avatar' => $link,
         ]);
 
-        return new MediaContentResource($mediaContent);
+        return new MediaContentResource(\Auth::user());
     }
 
     /**
